parking-lot-dash/requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
dash-daq>=0.5.0
paho-mqtt>=1.6.1
plotly>=5.17.0
pandas>=2.0.0
python-dotenv>=1.0.0
